moved language model specific fixtures to language models test folder
