+++ tests/test_modeling/test_model_choice.py
